// src/generated-api/tournament/models/tournament-bracket-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TournamentBracketRoundDto } from './tournament-bracket-round-dto';

/**
 * 
 * @export
 * @interface TournamentBracketDto
 */
export interface TournamentBracketDto {
    /**
     * 
     * @type {string}
     * @memberof TournamentBracketDto
     */
    type: TournamentBracketDtoTypeEnum;
    /**
     * 
     * @type {Array<TournamentBracketRoundDto>}
     * @memberof TournamentBracketDto
     */
    winning: Array<TournamentBracketRoundDto>;
    /**
     * 
     * @type {Array<TournamentBracketRoundDto>}
     * @memberof TournamentBracketDto
     */
    losing: Array<TournamentBracketRoundDto>;
}

/**
    * @export
    * @enum {string}
    */
export enum TournamentBracketDtoTypeEnum {
    SINGLEELIMINATION = 'SINGLE_ELIMINATION',
    DOUBLEELIMINATION = 'DOUBLE_ELIMINATION'
}