// src/generated-api/tournament/models/tournament-bracket-round-dto.ts
// tslint:disable
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { TournamentSeedDto } from './tournament-seed-dto';

/**
 * 
 * @export
 * @interface TournamentBracketRoundDto
 */
export interface TournamentBracketRoundDto {
    /**
     * 
     * @type {string}
     * @memberof TournamentBracketRoundDto
     */
    title: string;
    /**
     * 
     * @type {number}
     * @memberof TournamentBracketRoundDto
     */
    round: number;
    /**
     * 
     * @type {Array<TournamentSeedDto>}
     * @memberof TournamentBracketRoundDto
     */
    seeds: Array<TournamentSeedDto>;
}